+++ utils/src/main/lombok/de/larssh/utils/maven/package-info.java
@@ -1,5 +1,0 @@
-/**
- * Utilities for Maven plugins
- */
-@de.larssh.utils.annotations.NonNullByDefault
-package de.larssh.utils.maven;